+++ templates/example.js
@@ -1,3 +1,5 @@
+const outdent = require('outdent');
+
 module.exports = [{
   // only needed in case of `exampleFileLinkage: true`:
   name: ({ exampleFileLinkage }) =>
@@ -119,7 +121,7 @@ module.exports = [{
   // metro.config.js workarounds needed in case of `exampleFileLinkage: false`:
   name: ({ exampleName, exampleFileLinkage }) =>
     exampleFileLinkage ? undefined : `${exampleName}/metro.config.js`,
-  content: ({ moduleName, exampleName }) => `// metro.config.js
+  content: ({ packageName, exampleName }) => `// metro.config.js
 //
 // with multiple workarounds for this issue with symlinks:
 // https://github.com/facebook/metro/issues/1
@@ -145,13 +147,15 @@ module.exports = {
   },
 
   // quick workaround for another issue with symlinks
-  watchFolders: ['.', '..']
+  watchFolders: [path.resolve('.'), path.resolve('..')]
 }
 `,
 }, {
   name: ({ exampleName, writeExamplePodfile }) =>
     writeExamplePodfile ? `${exampleName}/ios/Podfile` : undefined,
-  content: ({ moduleName, exampleName }) => `platform :ios, '10.0'
+  content: ({ packageName, exampleName }) =>
+    outdent({ trimTrailingNewline: false })`
+	platform :ios, '10.0'
 
 	target '${exampleName}' do
 		rn_path = '../node_modules/react-native'
@@ -180,12 +184,12 @@ module.exports = {
 			'DevSupport'
 		]
 
-		pod '${moduleName}', :path => '../../${moduleName}.podspec'
+		pod '${packageName}', :path => '../../${packageName}.podspec'
 	end
 `,
 }, {
   name: ({ exampleName }) => `${exampleName}/App.js`,
-  content: ({ moduleName, objectClassName, view }) =>
+  content: ({ packageName, objectClassName, isView }) =>
     `/**
  * Sample React Native App
  *
@@ -198,8 +202,8 @@ module.exports = {
 
 import React, { Component } from 'react';
 import { Platform, StyleSheet, Text, View } from 'react-native';
-import ${objectClassName} from '${moduleName}';` +
-    (!view
+import ${objectClassName} from '${packageName}';` +
+    (!isView
       ? `
 
 export default class App extends Component<{}> {