+++ templates/general.js
@@ -1,19 +1,19 @@
 module.exports = [{
   name: () => 'README.md',
-  content: ({ moduleName, objectClassName }) =>
-    `# ${moduleName}
+  content: ({ packageName, objectClassName }) =>
+    `# ${packageName}
 
 ## Getting started
 
-\`$ npm install ${moduleName} --save\`
+\`$ npm install ${packageName} --save\`
 
 ### Mostly automatic installation
 
-\`$ react-native link ${moduleName}\`
+\`$ react-native link ${packageName}\`
 
 ## Usage
 \`\`\`javascript
-import ${objectClassName} from '${moduleName}';
+import ${objectClassName} from '${packageName}';
 
 // TODO: What to do with the module?
 ${objectClassName};
@@ -21,7 +21,7 @@ ${objectClassName};
 `,
 }, {
   name: () => 'package.json',
-  content: ({ moduleName, platforms, githubAccount, authorName, authorEmail, license }) => {
+  content: ({ packageName, platforms, githubAccount, authorName, authorEmail, license }) => {
     const files =
       `[
     "README.md",` +
@@ -30,12 +30,12 @@ ${objectClassName};
     "index.js"` +
     (platforms.indexOf('ios') >= 0 ? `,
     "ios",
-    "${moduleName}.podspec"` : ``) + `
+    "${packageName}.podspec"` : ``) + `
   ]`;
 
     const peerDependencies =
       `{
-    "react": "^16.8.1",
+    "react": ">=16.8.1",
     "react-native": ">=0.60.0-rc.0 <1.0.x"
   }`;
 
@@ -46,8 +46,8 @@ ${objectClassName};
   }`;
 
     return `{
-  "name": "${moduleName}",
-  "title": "${moduleName.split('-').map(word => word[0].toUpperCase() + word.substr(1)).join(' ')}",
+  "name": "${packageName}",
+  "title": "${packageName.split('-').map(word => word[0].toUpperCase() + word.substr(1)).join(' ')}",
   "version": "1.0.0",
   "description": "TODO",
   "main": "index.js",
@@ -57,8 +57,8 @@ ${objectClassName};
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/${githubAccount}/${moduleName}.git",
-    "baseUrl": "https://github.com/${githubAccount}/${moduleName}"
+    "url": "git+https://github.com/${githubAccount}/${packageName}.git",
+    "baseUrl": "https://github.com/${githubAccount}/${packageName}"
   },
   "keywords": [
     "react-native"
@@ -74,12 +74,14 @@ ${objectClassName};
   "devDependencies": ${devDependencies}
 }
 `;
-  },
+  }
 }, {
   // for module without view:
-  name: ({ view }) => !view && 'index.js',
+  name: ({ isView }) => !isView && 'index.js',
   content: ({ objectClassName }) =>
-    `import { NativeModules } from 'react-native';
+    `// main index.js
+
+import { NativeModules } from 'react-native';
 
 const { ${objectClassName} } = NativeModules;
 
@@ -87,9 +89,11 @@ export default ${objectClassName};
 `,
 }, {
   // for module with view:
-  name: ({ view }) => view && 'index.js',
+  name: ({ isView }) => isView && 'index.js',
   content: ({ objectClassName }) =>
-    `import { requireNativeComponent } from 'react-native';
+    `// main index.js
+
+import { requireNativeComponent } from 'react-native';
 
 const ${objectClassName} = requireNativeComponent('${objectClassName}', null);