+++ tests/with-injection/create/view/with-example/with-options/create-view-with-example-with-options.test.js
@@ -14,10 +14,10 @@ test('create alice-bobbi view module with example, with custom config options', 
     authorName: 'Alice',
     authorEmail: 'contact@alice.me',
     license: 'ISC',
-    view: true,
+    isView: true,
     generateExample: true,
     exampleName: 'test-demo',
-    exampleReactNativeVersion: 'react-native@0.60',
+    exampleReactNativeTemplate: 'react-native@0.60',
   };
 
   return lib(options, inject)