+++ tests/with-injection/create/with-defaults/no-prefix-needed/create-with-no-prefix-needed.test.js
@@ -2,19 +2,13 @@ const lib = require('../../../../../lib/lib.js');
 
 const ioInject = require('../../../helpers/io-inject.js');
 
-test('create alice-bobbi module with config options for Android only', () => {
+test('create with defaults, with no prefix needed', () => {
   const mysnap = [];
 
   const inject = ioInject(mysnap);
 
   const options = {
-    platforms: ['android'],
-    name: 'alice-bobbi',
-    nativePackageId: 'com.alicebits',
-    githubAccount: 'alicebits',
-    authorName: 'Alice',
-    authorEmail: 'contact@alice.me',
-    license: 'ISC',
+    name: 'react-native-alice',
   };
 
   return lib(options, inject)