+++ tests/with-mocks/cli/command/action-func/with-logging/with-bogus-platforms/empty-string/cli-command-with-empty-platforms-string.test.js
@@ -16,22 +16,22 @@ jest.mock('fs-extra', () => ({
     return Promise.resolve();
   },
 }));
-
-// TBD hackish mock:
-global.console = {
+jest.mock('console', () => ({
   info: (...args) => {
-    mockpushit({ info: [].concat(args) });
-  },
-  log: (...args) => {
     mockpushit({
       // TBD EXTRA WORKAROUND HACK for non-deterministic elapsed time in log
-      log: args.map(line => line.replace(/It took.*s/g, 'It took XXX'))
+      info: args.map(line => line.replace(/It took.*s/g, 'It took XXX'))
     });
   },
+  // console.log is no longer expected
+  // log: (...args) => ...
   warn: (...args) => {
     mockpushit({ warn: [].concat(args) });
   },
-};
+  error: (...args) => {
+    mockpushit({ error: [].concat(args) });
+  },
+}));
 
 test(`create alice-bobbi module with logging, with platforms: ''`, async () => {
   const args = ['alice-bobbi'];