+++ tests/with-mocks/lib/create/with-example/with-logging/with-options/create-with-example-with-options.test.js
@@ -1,42 +1,49 @@
 const lib = require('../../../../../../../lib/lib.js');
 
+const mockcwd = require('process').cwd();
+
 // special compact mocks for this test:
 const mysnap = [];
 const mockpushit = x => mysnap.push(x);
 jest.mock('fs-extra', () => ({
   outputFile: (outputFileName, theContent) => {
     mockpushit({
-      outputFileName: outputFileName.replace(/\\/g, '/'),
+      outputFileName: outputFileName.replace(mockcwd, '...').replace(/\\/g, '/'),
       theContent
     });
     return Promise.resolve();
   },
   ensureDir: (dir) => {
-    mockpushit({ ensureDir: dir.replace(/\\/g, '/') });
+    mockpushit({ ensureDir: dir.replace(mockcwd, '...').replace(/\\/g, '/') });
     return Promise.resolve();
   },
   readFileSync: (path) => {
-    mockpushit({ readFileSyncFromPath: path.replace(/\\/g, '/') });
+    mockpushit({ readFileSyncFromPath: path.replace(mockcwd, '...').replace(/\\/g, '/') });
     return `{ "name": "x", "scripts": { "test": "exit 1" } }`;
   },
   writeFileSync: (path, json, options) => {
     mockpushit({
-      writeFileSyncToPath: path.replace(/\\/g, '/'),
+      writeFileSyncToPath: path.replace(mockcwd, '...').replace(/\\/g, '/'),
       json,
       options
     });
   },
 }));
 jest.mock('execa', () => ({
-  commandSync: (command, options) => {
+  commandSync: (command, opts) => {
+    const options = { ...opts, ...(opts.cwd ? { cwd: opts.cwd.replace(mockcwd, '...').replace(/\\/g, '/') } : {}) };
     mockpushit({ commandSync: command, options });
   }
 }));
-
-// TBD hackish mock:
-global.console = {
+jest.mock('react-native-init-func', () => (projectNameArray, opts) => {
+  const options = { ...opts, ...(opts.directory ? { directory: opts.directory.replace(/\\/g, '/') } : {}) };
+  mockpushit({ call: 'reactNativeInit', nameArray: projectNameArray, options });
+});
+jest.mock('console', () => ({
   info: (...args) => {
-    mockpushit({ info: [].concat(args) });
+    mockpushit({
+      info: args.map(line => line.replace(mockcwd, '...').replace(/\\/g, '/'))
+    });
   },
   log: (...args) => {
     mockpushit({ log: [].concat(args) });
@@ -44,14 +51,16 @@ global.console = {
   warn: (...args) => {
     mockpushit({ warn: [].concat(args) });
   },
-};
+  error: (...args) => {
+    mockpushit({ error: [].concat(args) });
+  },
+}));
 
 test('create alice-bobbi module using mocked lib with logging, with example, for Android & iOS with config options including `exampleFileLinkage: true`', async () => {
   const options = {
     platforms: ['android', 'ios'],
     name: 'alice-bobbi',
-    prefix: 'ABC',
-    packageIdentifier: 'com.alicebits',
+    nativePackageId: 'com.alicebits',
     tvosEnabled: true,
     githubAccount: 'alicebits',
     authorName: 'Alice',
@@ -60,7 +69,7 @@ test('create alice-bobbi module using mocked lib with logging, with example, for
     generateExample: true,
     exampleFileLinkage: true,
     exampleName: 'demo',
-    exampleReactNativeVersion: 'react-native@npm:react-native-tvos'
+    exampleReactNativeTemplate: 'react-native@npm:react-native-tvos'
   };
 
   await lib(options);